[MOD]use handler to refresh ui

+++ app/src/main/java/com/echo/timewidget/TimeWidgetProvider.java
@@ -1,9 +1,7 @@
 package com.echo.timewidget;
 
-import android.app.Activity;
 import android.appwidget.AppWidgetManager;
 import android.appwidget.AppWidgetProvider;
-import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -12,6 +10,7 @@ import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
 import android.os.Bundle;
 import android.os.Environment;
+import android.os.Message;
 import android.support.v8.renderscript.Allocation;
 import android.support.v8.renderscript.RenderScript;
 import android.support.v8.renderscript.ScriptIntrinsicBlur;
@@ -37,6 +36,9 @@ import java.util.Date;
  */
 public class TimeWidgetProvider extends AppWidgetProvider {
 
+    TimeWidgetHandler mHandler;
+
+
     public TimeWidgetProvider() {
         super();
     }
@@ -47,97 +49,87 @@ public class TimeWidgetProvider extends AppWidgetProvider {
 
         Log.d("life", "receive");
 
-        RemoteViews remoteViews = new RemoteViews(context.getPackageName(), R.layout.time_widget_layout);
-
-        updateTextView(remoteViews);
-
-        //获得appwidget管理实例，用于管理appwidget以便进行更新操作
-        AppWidgetManager appWidgetManager = AppWidgetManager.getInstance(context);
-
-        //相当于获得所有本程序创建的appwidget
-        ComponentName componentName = new ComponentName(context, TimeWidgetProvider.class);
-
-        //更新appwidget
-        appWidgetManager.updateAppWidget(componentName, remoteViews);
     }
 
     @Override
     public void onUpdate(final Context context, AppWidgetManager appWidgetManager, int[] appWidgetIds) {
         super.onUpdate(context, appWidgetManager, appWidgetIds);
 
+        if(mHandler==null){
+            mHandler=new TimeWidgetHandler(context);
+        }else{
+            mHandler.setContext(context);
+        }
+
         Log.d("life", "update");
-
-        RemoteViews remoteViews = new RemoteViews(context.getPackageName(), R.layout.time_widget_layout);
-
-        updateTextView(remoteViews);
 
         final SharedPreferences sharedPreferences = context.getSharedPreferences("echo", Context.MODE_PRIVATE);
         final int day = sharedPreferences.getInt("date", -1);
         final int current = Calendar.getInstance().get(Calendar.DATE);
 
-        if (current != day) {
 
-            new Thread() {
+        new Thread() {
 
-                @Override
-                public void run() {
-                    super.run();
+            @Override
+            public void run() {
+                super.run();
+                if (current != day) {
+
                     Log.d("Download", "start");
                     Bitmap bkg = GetLocalOrNetBitmap("http://www.dujin.org/sys/bing/1920.php");
 
                     if (bkg != null) {
-
+                        Log.d("bkg", "not null");
                         File file = new File(Environment.getExternalStorageDirectory().getPath() + "/wallpaper");
                         File newFile = new File(Environment.getExternalStorageDirectory().getPath() + "/newwallpaper");
+
+                        boolean getBitmapFlag = false;
 
                         try {
                             OutputStream os = new FileOutputStream(newFile);
                             bkg.compress(Bitmap.CompressFormat.JPEG, 100, os);
                             os.flush();
                             os.close();
+                            getBitmapFlag = true;
                         } catch (FileNotFoundException e) {
+                            Log.d("bkg", e.getMessage());
                             e.printStackTrace();
-                            return;
                         } catch (IOException e) {
+                            Log.d("bkg", e.getMessage());
                             e.printStackTrace();
-                            return;
                         }
 
-                        Log.d("GETBitmap", "ok");
-                        SharedPreferences.Editor editor = sharedPreferences.edit();
-                        editor.putInt("date", current);
-                        editor.putBoolean("flag", true);
-                        editor.apply();
+                        if (getBitmapFlag) {
+                            Log.d("GETBitmap", "ok");
+                            SharedPreferences.Editor editor = sharedPreferences.edit();
+                            editor.putInt("date", current);
+                            editor.apply();
 
-                        if (newFile.exists()) {
-                            if (file.exists()) {
-                                file.delete();
+                            if (newFile.exists()) {
+                                if (file.exists()) {
+                                    file.delete();
+                                }
+                                newFile.renameTo(file);
                             }
-                            newFile.renameTo(file);
                         }
+                    }else{
+                        Log.d("bkg", "null");
                     }
+
                 }
-            }.start();
-        }
+                Bitmap bitmap =doRender(context);
+                if(bitmap!=null){
+                    Log.d("bitmap", "not null");
+                    Message msg =mHandler.obtainMessage();
+                    msg.obj=bitmap;
+                    msg.what=TimeWidgetHandler.UPDATE_VIEW;
+                    msg.sendToTarget();
+                }else{
+                    Log.d("bitmap", "null");
+                }
+            }
+        }.start();
 
-        boolean flag = sharedPreferences.getBoolean("flag", true);
-
-        Log.d("updateView", "flag" + flag);
-
-        if (flag && updateBackView(context, remoteViews)) {
-
-            Log.d("updateView", "success");
-
-            SharedPreferences.Editor editor = sharedPreferences.edit();
-            editor.putBoolean("flag", false);
-            editor.apply();
-        }
-
-        //相当于获得所有本程序创建的appwidget
-        ComponentName componentName = new ComponentName(context, TimeWidgetProvider.class);
-
-        //更新appwidget
-        appWidgetManager.updateAppWidget(componentName, remoteViews);
     }
 
     @Override
@@ -166,24 +158,8 @@ public class TimeWidgetProvider extends AppWidgetProvider {
         super.onRestored(context, oldWidgetIds, newWidgetIds);
     }
 
-    private void updateTextView(RemoteViews remoteViews) {
-
-        //Time 1 February 12
-        //Time 2 Apr 12
-        //Time 3 Apr 16
-        //Time 4 Apr 24
-        Calendar calendar1 = Calendar.getInstance();
-        calendar1.set(2016, 1, 12);
-        Calendar calendar2 = Calendar.getInstance();
-        calendar2.set(2016, 3, 12);
-
-        int time1 = daysBetween(calendar1.getTime(), new Date());
-        int time2 = daysBetween(calendar2.getTime(), new Date());
-        remoteViews.setTextViewText(R.id.day1, time1 + 1 + "");
-        remoteViews.setTextViewText(R.id.day2, " & " + (time2 + 1));
-    }
-
     public static Bitmap GetLocalOrNetBitmap(String url) {
+        Log.d("download",url);
         Bitmap bitmap = null;
         InputStream in = null;
         BufferedOutputStream out = null;
@@ -197,6 +173,7 @@ public class TimeWidgetProvider extends AppWidgetProvider {
             bitmap = BitmapFactory.decodeByteArray(data, 0, data.length);
             return bitmap;
         } catch (Exception e) {
+            Log.d("download",e.getMessage());
             e.printStackTrace();
             return null;
         }
@@ -210,18 +187,21 @@ public class TimeWidgetProvider extends AppWidgetProvider {
         }
     }
 
-    private boolean updateBackView(Context context, RemoteViews remoteViews) {
 
-        Log.d("updateView", "start update");
+    private Bitmap doRender(Context mContext){
+
 
         Bitmap bkg = BitmapFactory.decodeFile(Environment.getExternalStorageDirectory().getPath() + "/wallpaper");
         if (bkg != null) {
+            bkg = bkg.copy(Bitmap.Config.ARGB_8888, true);
+        }else{
+            bkg = BitmapFactory.decodeResource(mContext.getResources(),R.drawable.sun_flower);
             bkg = bkg.copy(Bitmap.Config.ARGB_8888, true);
         }
 
         if (bkg == null) {
             Log.d("updateview", "empty");
-            return false;
+            return null;
         }
 
         Canvas canvas = new Canvas(bkg);
@@ -229,7 +209,7 @@ public class TimeWidgetProvider extends AppWidgetProvider {
         canvas.translate(0, 0);
         canvas.drawBitmap(bkg, 0, 0, null);
 
-        RenderScript rs = RenderScript.create(context);
+        RenderScript rs = RenderScript.create(mContext);
 
         Allocation overlayAlloc = Allocation.createFromBitmap(rs, bkg);
 
@@ -243,38 +223,12 @@ public class TimeWidgetProvider extends AppWidgetProvider {
 
         overlayAlloc.copyTo(bkg);
 
-        remoteViews.setBitmap(R.id.back, "setImageBitmap", bkg);
-
         rs.destroy();
 
-        Log.d("updateView", "update finish");
+        return bkg;
 
-        return true;
     }
 
-    /**
-     * 计算两个日期之间相差的天数
-     *
-     * @param smdate 较小的时间
-     * @param bdate 较大的时间
-     * @return 相差天数
-     * @throws ParseException
-     */
-    private int daysBetween(Date smdate, Date bdate) {
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-        try {
-            smdate = sdf.parse(sdf.format(smdate));
-            bdate = sdf.parse(sdf.format(bdate));
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
-        Calendar cal = Calendar.getInstance();
-        cal.setTime(smdate);
-        long time1 = cal.getTimeInMillis();
-        cal.setTime(bdate);
-        long time2 = cal.getTimeInMillis();
-        long between_days = (time2 - time1) / (1000 * 3600 * 24);
 
-        return Integer.parseInt(String.valueOf(between_days));
-    }
+
 }
